+++ src/Administration/Resources/app/administration/src/app/component/entity/sw-meteor-entity-data-table/sw-meteor-entity-data-table.spec/fixtures.ts
@@ -1,0 +1,414 @@
+/**
+ * @sw-package framework
+ */
+
+/* eslint-disable sw-deprecation-rules/private-feature-declarations, jest/require-top-level-describe */
+
+import { flushPromises, mount } from '@vue/test-utils';
+import type { VueWrapper } from '@vue/test-utils';
+import { h, nextTick, ref } from 'vue';
+import type { SetupContext } from 'vue';
+import { overrideComponentSetup, _overridesMap } from 'src/app/adapter/composition-extension-system';
+import type EntityCollection from '@shopware-ag/meteor-admin-sdk/es/_internals/data/EntityCollection';
+import type { ApiContext } from '@shopware-ag/meteor-admin-sdk/es/_internals/data/EntityCollection';
+import type CriteriaType from 'src/core/data/criteria.data';
+import type Repository from 'src/core/data/repository.data';
+import { isNativeShopwareComponentName } from 'src/app/component/native-shopware-components';
+import SwMeteorEntityDataTable from '../sw-meteor-entity-data-table.vue';
+import type {
+    SwMeteorEntityDataTableColumn,
+    SwMeteorEntityDataTableCriteriaResolver,
+    SwMeteorEntityDataTableCriteriaResolverPayload,
+    SwMeteorEntityDataTableLayout,
+    SwMeteorEntityDataTableState,
+} from '../sw-meteor-entity-data-table.types';
+import { MtDataTableStub, globalStubs } from './stubs';
+
+export const componentName = 'sw-meteor-entity-data-table';
+export const { Criteria } = Shopware.Data;
+export const shopwareApplication = Shopware.Application as unknown as {
+    view: {
+        router?: TestRouter;
+    };
+};
+
+export type TestRepository = Repository<keyof EntitySchema.Entities>;
+
+export type TestColumn = SwMeteorEntityDataTableColumn;
+
+export type TestRecord = {
+    id: string;
+    name: string;
+    [key: string]: unknown;
+};
+
+export type TestSearchMock = jest.Mock<Promise<EntityCollection<keyof EntitySchema.Entities>>, [CriteriaType, ApiContext]>;
+export type TestDeleteMock = jest.Mock<Promise<unknown>, [string, ApiContext]>;
+export type TestSyncDeletedMock = jest.Mock<Promise<void>, [string[], ApiContext]>;
+export type TestSaveMock = jest.Mock<Promise<void>, [TestRecord, ApiContext]>;
+
+export type TestProps = {
+    repository: TestRepository;
+    columns: TestColumn[];
+    identifier?: string;
+    criteria?: CriteriaType | null;
+    criteriaResolver?: SwMeteorEntityDataTableCriteriaResolver | null;
+    context?: ApiContext | null;
+    initialPage?: number;
+    initialLimit?: number;
+    initialSearchTerm?: string;
+    initialSort?: SwMeteorEntityDataTableState['sort'] | null;
+    paginationOptions?: number[];
+    layout?: SwMeteorEntityDataTableLayout;
+    searchable?: boolean;
+    reloadable?: boolean;
+    selectable?: boolean;
+    detailRoute?: string;
+    allowEdit?: boolean;
+    allowInlineEdit?: boolean;
+    allowDelete?: boolean;
+    hideTableSettings?: boolean;
+    additionalContextButtons?: Array<{
+        key: string;
+        label: string;
+        type?: 'default' | 'active' | 'critical';
+    }>;
+};
+
+export type SlotRenderer = string | ((slotProps: Record<string, unknown>) => ReturnType<typeof h>);
+export type SlotRenderers = Record<string, SlotRenderer>;
+
+export type TestRouter = {
+    push: jest.Mock;
+};
+
+export type TestUserConfigEntity = {
+    id?: string;
+    key?: string;
+    userId?: string;
+    value?: unknown;
+};
+
+export type TestUserConfigRepository = {
+    search: jest.Mock<Promise<TestUserConfigEntity[]>, [CriteriaType, ApiContext]>;
+    save: jest.Mock<Promise<void>, [TestUserConfigEntity, ApiContext]>;
+    create: jest.Mock<TestUserConfigEntity, [ApiContext]>;
+    getStoredEntity: () => TestUserConfigEntity | null;
+};
+
+export const columns: TestColumn[] = [
+    {
+        label: 'Name',
+        property: 'name',
+    },
+];
+
+export const inlineEditColumns: TestColumn[] = [
+    {
+        label: 'Name',
+        property: 'name',
+        inlineEdit: 'string',
+    },
+];
+
+export const persistedSettingsColumns: TestColumn[] = [
+    {
+        label: 'Name',
+        property: 'name',
+    },
+    {
+        label: 'Link',
+        property: 'link',
+    },
+];
+
+export const records: TestRecord[] = [
+    {
+        id: 'record-1',
+        name: 'First record',
+    },
+    {
+        id: 'record-2',
+        name: 'Second record',
+    },
+];
+
+export function createSearchResult(
+    resultRecords: TestRecord[] = records,
+    total = resultRecords.length,
+): EntityCollection<keyof EntitySchema.Entities> {
+    return Object.assign([...resultRecords], { total }) as unknown as EntityCollection<keyof EntitySchema.Entities>;
+}
+
+export function createRepositoryMock(searchResult = createSearchResult()): TestRepository {
+    const search: TestSearchMock = jest.fn<ReturnType<TestSearchMock>, Parameters<TestSearchMock>>();
+    const deleteMock: TestDeleteMock = jest.fn<ReturnType<TestDeleteMock>, Parameters<TestDeleteMock>>();
+    const syncDeletedMock: TestSyncDeletedMock = jest.fn<ReturnType<TestSyncDeletedMock>, Parameters<TestSyncDeletedMock>>();
+    const saveMock: TestSaveMock = jest.fn<ReturnType<TestSaveMock>, Parameters<TestSaveMock>>();
+    search.mockResolvedValue(searchResult);
+    deleteMock.mockResolvedValue({});
+    syncDeletedMock.mockResolvedValue();
+    saveMock.mockResolvedValue();
+
+    return {
+        search,
+        delete: deleteMock,
+        syncDeleted: syncDeletedMock,
+        save: saveMock,
+    } as unknown as TestRepository;
+}
+
+export function createRepositoryMockWithSearch(search: TestSearchMock): TestRepository {
+    const deleteMock: TestDeleteMock = jest.fn<ReturnType<TestDeleteMock>, Parameters<TestDeleteMock>>();
+    const syncDeletedMock: TestSyncDeletedMock = jest.fn<ReturnType<TestSyncDeletedMock>, Parameters<TestSyncDeletedMock>>();
+    const saveMock: TestSaveMock = jest.fn<ReturnType<TestSaveMock>, Parameters<TestSaveMock>>();
+    deleteMock.mockResolvedValue({});
+    syncDeletedMock.mockResolvedValue();
+    saveMock.mockResolvedValue();
+
+    return {
+        search,
+        delete: deleteMock,
+        syncDeleted: syncDeletedMock,
+        save: saveMock,
+    } as unknown as TestRepository;
+}
+
+export function createUserConfigRepositoryMock(initialEntity: TestUserConfigEntity | null = null): TestUserConfigRepository {
+    let storedEntity = initialEntity;
+    const search = jest.fn<Promise<TestUserConfigEntity[]>, [CriteriaType, ApiContext]>(() =>
+        Promise.resolve(storedEntity ? [storedEntity] : []),
+    );
+    const save = jest.fn<Promise<void>, [TestUserConfigEntity, ApiContext]>((entity) => {
+        storedEntity = entity;
+
+        return Promise.resolve();
+    });
+    const create = jest.fn<TestUserConfigEntity, [ApiContext]>(() => ({
+        id: 'created-user-config-id',
+    }));
+
+    return {
+        search,
+        save,
+        create,
+        getStoredEntity: () => storedEntity,
+    };
+}
+
+export function mockUserConfigRepository(userConfigRepository: TestUserConfigRepository): void {
+    const aclService = Shopware.Service('acl') as {
+        can: (privilege: string) => boolean;
+    };
+    const repositoryFactory = Shopware.Service('repositoryFactory') as {
+        create: (entityName: keyof EntitySchema.Entities) => Repository<keyof EntitySchema.Entities>;
+    };
+
+    jest.spyOn(aclService, 'can').mockReturnValue(true);
+    jest.spyOn(repositoryFactory, 'create').mockImplementation((entityName) => {
+        if (entityName === 'user_config') {
+            return userConfigRepository as unknown as Repository<keyof EntitySchema.Entities>;
+        }
+
+        return createRepositoryMock() as unknown as Repository<keyof EntitySchema.Entities>;
+    });
+}
+
+export function setCurrentUserWithUserConfigPrivileges(): void {
+    Shopware.Store.get('session').setCurrentUser({
+        id: 'current-user-id',
+        admin: true,
+        aclRoles: [
+            {
+                privileges: [
+                    'user_config:read',
+                    'user_config:create',
+                    'user_config:update',
+                ],
+            },
+        ],
+    } as EntitySchema.user);
+}
+
+export function getSearchMock(repository: TestRepository): TestSearchMock {
+    return (repository as unknown as { search: TestSearchMock }).search;
+}
+
+export function getDeleteMock(repository: TestRepository): TestDeleteMock {
+    return (repository as unknown as { delete: TestDeleteMock }).delete;
+}
+
+export function getSyncDeletedMock(repository: TestRepository): TestSyncDeletedMock {
+    return (repository as unknown as { syncDeleted: TestSyncDeletedMock }).syncDeleted;
+}
+
+export function getSaveMock(repository: TestRepository): TestSaveMock {
+    return (repository as unknown as { save: TestSaveMock }).save;
+}
+
+export function createWrapper(
+    options: {
+        props?: Partial<TestProps>;
+        slots?: SlotRenderers;
+        router?: TestRouter;
+    } = {},
+) {
+    return mount(SwMeteorEntityDataTable, {
+        props: {
+            repository: createRepositoryMock(),
+            columns,
+            ...options.props,
+        },
+        slots: options.slots,
+        global: {
+            mocks: options.router
+                ? {
+                      $router: options.router,
+                  }
+                : {},
+            stubs: globalStubs,
+        },
+    });
+}
+
+export function createListingConsumerWrapper(repository = createRepositoryMock()) {
+    return mount(
+        {
+            name: 'sw-meteor-entity-data-table-listing-consumer',
+            components: {
+                SwMeteorEntityDataTable,
+            },
+            mixins: [
+                Shopware.Mixin.getByName('listing'),
+            ],
+            template: `
+                <sw-meteor-entity-data-table
+                    :repository="repository"
+                    :columns="columns"
+                    selectable
+                    @selection-change="updateSelection"
+                />
+            `,
+            data() {
+                return {
+                    repository,
+                    columns,
+                    disableRouteParams: true,
+                };
+            },
+            methods: {
+                getList: jest.fn(),
+            },
+        },
+        {
+            global: {
+                mocks: {
+                    $route: {
+                        name: 'sw.product.index',
+                        query: {},
+                        params: {},
+                    },
+                    $router: {
+                        push: jest.fn(),
+                        replace: jest.fn(),
+                    },
+                },
+                provide: {
+                    searchRankingService: {
+                        isValidTerm: (term?: string) => {
+                            return !!term && term.trim().length >= 1;
+                        },
+                        getSearchFieldsByEntity: jest.fn(),
+                        buildSearchQueriesForEntity: jest.fn(),
+                    },
+                    feature: {},
+                },
+                stubs: globalStubs,
+            },
+        },
+    );
+}
+
+export function getTable(wrapper: VueWrapper) {
+    return wrapper.findComponent(MtDataTableStub);
+}
+
+export function getLastSearchCriteria(repository: TestRepository): CriteriaType {
+    const searchMock = getSearchMock(repository);
+    const lastCall = searchMock.mock.calls[searchMock.mock.calls.length - 1];
+
+    expect(lastCall).toBeDefined();
+
+    return lastCall[0];
+}
+
+export function getSetupState(wrapper: VueWrapper): Record<string, unknown> {
+    return (wrapper.vm.$ as unknown as { setupState: Record<string, unknown> }).setupState;
+}
+
+export function createDeferred<TValue>() {
+    let resolve: (value: TValue) => void = () => {};
+    let reject: (reason?: unknown) => void = () => {};
+    const promise = new Promise<TValue>((promiseResolve, promiseReject) => {
+        resolve = promiseResolve;
+        reject = promiseReject;
+    });
+
+    return {
+        promise,
+        resolve,
+        reject,
+    };
+}
+
+export {
+    SwMeteorEntityDataTable,
+    flushPromises,
+    h,
+    isNativeShopwareComponentName,
+    mount,
+    MtDataTableStub,
+    nextTick,
+    overrideComponentSetup,
+    ref,
+    globalStubs,
+    _overridesMap,
+};
+
+export type {
+    ApiContext,
+    CriteriaType,
+    EntityCollection,
+    Repository,
+    SetupContext,
+    SwMeteorEntityDataTableColumn,
+    SwMeteorEntityDataTableCriteriaResolver,
+    SwMeteorEntityDataTableCriteriaResolverPayload,
+    SwMeteorEntityDataTableLayout,
+    SwMeteorEntityDataTableState,
+    VueWrapper,
+};
+
+export function registerSwMeteorEntityDataTableHooks(): void {
+    const originalRouter = shopwareApplication.view.router;
+    const originalCurrentUser = Shopware.Store.get('session').currentUser;
+
+    beforeEach(() => {
+        delete _overridesMap[componentName];
+        Shopware.Component.getOverrideRegistry().delete(componentName);
+    });
+
+    afterEach(() => {
+        shopwareApplication.view.router = originalRouter;
+        Shopware.Store.get('shopwareApps').selectedIds = [];
+        Shopware.Store.get('swBulkEdit').selectedIds = [];
+
+        if (originalCurrentUser) {
+            Shopware.Store.get('session').setCurrentUser(originalCurrentUser);
+        } else {
+            Shopware.Store.get('session').removeCurrentUser();
+        }
+
+        jest.restoreAllMocks();
+    });
+}
